Limit layout flickering on page load
The Tuleap project sidebar component comes with a stylesheet to limit
the layout flickering. The stylesheet sets a width and height on the
`tuleap-project-sidebar` HTML element when it is not yet defined.
This way when the project sidebar code is loaded the layout does not
move.

We also load the latest version of @tuleap/project-sidebar to be
consistent with the rest of Tuleap.

Change-Id: I2235a0759bbe04bfa2226afca66702208f424f93