Avoid loading Font Awesome entirely for one icon
Font Awesome is quite large (100 KB+) and the minimal usage of the
provided icons in the theme does not justify its presence.

This contribution replaces Font Awesome with a minimal font including
only the icon we need.

The WOFF2 file is inlined directly in the CSS file since it will be
requested on all pages anyway.

The font has been built using https://icones.js.org/

Change-Id: I929339a9a8ae253a99b49c6b7c306f66f220ca8f